requirements/base.txt
django-mptt
psycopg2
django >= 2.0
